+++ src/reporter/query_1T1ENA.py
@@ -99,14 +99,12 @@ def query_1T1ENA(entity_id,   # In Path
 
         index = [] if aggr_method and not aggr_period else entities[0]['index']
         res = {
-            'entityId': entity_id,
-            'entityType': entities[0]['type'],
+            'id': entity_id,
+            'type': entities[0]['type'],
             'index': index,
             'attributes': attributes
         }
         logging.getLogger(__name__).info("Query processed successfully")
-        logging.warning(
-            "usage of id and type rather than entityId and entityType from version 0.9")
         return res
 
     r = []
@@ -117,8 +115,6 @@ def query_1T1ENA(entity_id,   # In Path
 def query_1T1ENA_value(*args, **kwargs):
     res = query_1T1ENA(*args, **kwargs)
     if isinstance(res, dict):
-        res.pop('entityId', None)
-        res.pop('entityType', None)
-    logging.warning(
-        "usage of id and type rather than entityId and entityType from version 0.9")
+        res.pop('id', None)
+        res.pop('type', None)
     return res