+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 ### Bug fixes
 
 - Fix to return OK/200 and an empty result set when no data found (#720)
+- Fix to return id and type in place of entityId and entityType (#751)
 
 ### Continuous Integration